Remove NetworkException and NetworkError
NetworkException was a generic exception that got thrown if HttpClient.GetAsync threw an exception. But since
the only exception it can throw is if the arguments are null, it is now replaced by prior argument null checking.
NetworkError was returned by command handlers if a NetworkException occurred, so it is useless without
NetworkException.