Remove different QueryParameterFields Classes in favor of one BoardFields Class
It seems that the fields parameter for each entity is always the same. Only how they are queried differs,
namely with which query name (?queryName=queryValue).
Therefore it vastly reduces code duplication if there is just one EntityFields class for each entity type.
The QueryParameter classes are responsible for putting together the query string, so they can decide
which query name to use, because that is different from query to query.