Improve Action Entity Class
It is now abstract, because there are many types of actions. Each type has different layout of data sent
with it. In the future there will be one inherited class for each type.
For the same reason, the Data and Type properties are removed.
The property and class ActionExecutor was added, because some properties of the member who
did the action may be sent with it in the same response.